+++ RELEASES.md
@@ -31,6 +31,7 @@
 17. improve memory defragmentation (v0.5.0b 2025-01-24)
 18. 2x improve memory defragmentation (v0.5.0b 2025-01-26)
 19. improve load speed and memory use  (v0.5.0b 2025-01-27)
+20. add firstKey (min) and lastKey (max) functions  (v0.5.0b 2025-01-27)
 
 ## Fixes
 1. fix conversions (v0.4.3.3b 2025-01-05)